+++ hardhat.config.zksync.ts
@@ -1,0 +1,145 @@
+import "module-alias/register";
+
+import "@matterlabs/hardhat-zksync";
+import "@matterlabs/hardhat-zksync-solc";
+import "@matterlabs/hardhat-zksync-verify";
+import "@nomicfoundation/hardhat-chai-matchers";
+import "@nomiclabs/hardhat-ethers";
+import * as dotenv from "dotenv";
+import "hardhat-dependency-compiler";
+import "hardhat-deploy";
+import "hardhat-gas-reporter";
+import { HardhatUserConfig, extendConfig, task } from "hardhat/config";
+import { HardhatConfig } from "hardhat/types";
+import "solidity-docgen";
+
+dotenv.config();
+
+const DEPLOYER_PRIVATE_KEY = process.env.DEPLOYER_PRIVATE_KEY;
+
+extendConfig((config: HardhatConfig) => {
+  if (process.env.EXPORT !== "true") {
+    config.external = {
+      ...config.external,
+      deployments: {
+        zksyncsepolia: ["node_modules/@venusprotocol/protocol-reserve/deployments/zksyncsepolia"],
+      },
+    };
+  }
+});
+
+task("accounts", "Prints the list of accounts", async (taskArgs, hre) => {
+  const accounts = await hre.ethers.getSigners();
+
+  for (const account of accounts) {
+    console.log(account.address);
+  }
+});
+
+const config: HardhatUserConfig = {
+  defaultNetwork: "hardhat",
+  zksolc: {
+    version: "1.5.0",
+  },
+  solidity: {
+    compilers: [
+      {
+        version: "0.8.25",
+        settings: {
+          optimizer: {
+            enabled: true,
+            details: {
+              yul: !process.env.CI,
+            },
+          },
+          evmVersion: "paris",
+          outputSelection: {
+            "*": {
+              "*": ["storageLayout"],
+            },
+          },
+        },
+      },
+      {
+        version: "0.6.6",
+        settings: {
+          optimizer: {
+            enabled: true,
+            details: {
+              yul: !process.env.CI,
+            },
+          },
+          outputSelection: {
+            "*": {
+              "*": ["storageLayout"],
+            },
+          },
+        },
+      },
+      {
+        version: "0.5.16",
+        settings: {
+          optimizer: {
+            enabled: true,
+            details: {
+              yul: !process.env.CI,
+            },
+          },
+          outputSelection: {
+            "*": {
+              "*": ["storageLayout"],
+            },
+          },
+        },
+      },
+    ],
+  },
+  networks: {
+    hardhat: {
+      allowUnlimitedContractSize: true,
+      loggingEnabled: false,
+      live: false,
+      zksync: true,
+    },
+    zksyncsepolia: {
+      url: process.env.ARCHIVE_NODE_zksyncsepolia || "https://sepolia.era.zksync.dev",
+      ethNetwork: "sepolia",
+      verifyURL: "https://explorer.sepolia.era.zksync.dev/contract_verification",
+      accounts: DEPLOYER_PRIVATE_KEY ? [`0x${DEPLOYER_PRIVATE_KEY}`] : [],
+      zksync: true,
+      live: true,
+    },
+  },
+  gasReporter: {
+    enabled: process.env.REPORT_GAS !== undefined,
+    currency: "USD",
+  },
+  paths: {
+    tests: "./tests",
+    cache: "./cache-zk",
+    artifacts: "./artifacts-zk",
+  },
+  docgen: {
+    outputDir: "./docs",
+    pages: "files",
+    templates: "./docgen-templates",
+  },
+  // Hardhat deploy
+  namedAccounts: {
+    deployer: {
+      default: 0, // here this will by default take the first account as deployer
+      acc1: 1,
+      acc2: 2,
+      proxyAdmin: 3,
+      acc3: 4,
+    },
+  },
+  dependencyCompiler: {
+    paths: [
+      "hardhat-deploy/solc_0.8/proxy/OptimizedTransparentUpgradeableProxy.sol",
+      "hardhat-deploy/solc_0.8/openzeppelin/proxy/transparent/ProxyAdmin.sol",
+    ],
+  },
+};
+
+export default config;